Split up constants and document helpers.
