+++ server/rules/homestead/read.test.ts
@@ -1,16 +1,16 @@
 import { assertFails, assertSucceeds } from '@firebase/rules-unit-testing';
-import { Firestore } from '@test-helpers/types';
-
+import { Collections } from '@test-helpers/constants';
 import {
-  COLLECTIONS,
   generateId,
   generateSecurityRecordAny,
   generateUserId,
   membershipPath,
-} from '../../test-helpers/constants';
+} from '@test-helpers/documents';
+import { Firestore } from '@test-helpers/types';
+
 import { setup, teardown } from '../../test-helpers/firestore-helpers';
 
-const COLLECTION = COLLECTIONS.HOMESTEADS;
+const COLLECTION = Collections.Homesteads;
 const DOC_ID_1 = generateId();
 const DOC_ID_2 = generateId();
 const USER_ID = generateUserId();
