+++ server/types.ts
@@ -1,0 +1,1 @@
+type Enum<T> = T[keyof T];
